+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Release History
 
-## 12.7.0-preview.2 (Unreleased)
+## 12.7.0 (2020-11-10)
+- Includes all features from 12.7.0-preview.1
 - Fixed bug where BlobContainerClient.SetAccessPolicy() would throw an exception if signed identifier permissions were not in the correct order.
 - Added seekability to BaseBlobClient.OpenRead().
 - Added additional info to exception messages.